gluster:Remove support for Replace gluster brick with data migration.
This patch removes verbs related to replace gluster brick
with data migration. Now only replace brick with commit
is retained. As this feature is deprecated in
glusterfs-3.7 on wards. These verbs were not consumed
by engine, so it should not cause any backward-compatibility
issues when removed. The list of verbs removed are as
follows:
  *volumeReplaceBrickStart
  *volumeReplaceBrickAbort
  *volumeReplaceBrickPause
  *volumeReplaceBrickStatus
  *volumeReplaceBrickCommit

Change-Id: I1569972120da27c17407bec5301b1ba417ce0768
Signed-off-by: Darshan N <dnarayan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40545
Continuous-Integration: Jenkins CI
Reviewed-by: Bala.FA <barumuga@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>