net: Force blocking DHCP when restoring networks
This is done right before restoring the network configuration, and
before calling setupNetworks. It forces the configurator to wait for
an IP address to be configured on the devices before restoration is
completed. This prevents VDSM from possibly report missing IP
addresses on interfaces that had been restored right before it was
started.

Change-Id: Ibc8b84a82794ac97eba6d34ec9d54430e387b659
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1218637
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39381
Reviewed-on: https://gerrit.ovirt.org/39529
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>