resize-lun: Support guids parameter in getDeviceList
Add optional "guids" parameter to getDeviceList. If specified,
getDevicesList will get info and test only these devices.

In the resize LUN flow, we need to rescan and resize the devices before
resizing the PV. Engine will invoke getDeviceList verb with specific
LUNs ids.

In getDeviceList we test PV creation and this is a heavyweight
operation, using the guids parameter we limit the operation to the
specified devices.

Examples:

    getDeviceList
        return all devices

    getDeviceList FCP
        return only FCP devices

    getDeviceList ISCSI guid1 guid2
        return info for guid1 and guid2, assuming ISCSI type

Change-Id: Ic173d94a132e617ae97353d38520a86bede657d7
Bug-Url:https://bugzilla.redhat.com/609689
Signed-off-by: Fred Rolland <frolland@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40661
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Tested-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Jenkins CI
Continuous-Integration: Jenkins CI
Tested-by: Jenkins CI
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>