net: remove support for sriov hook
Now that vdsm supports sriov natively, it does not need a special
hook for this. Moreover, allowing this hook to be installed might
introduce bugs originating from the use of the two methods at the
same time. vdsm now cannot be installed if hook-sriov package is
already installed on the system.

Change-Id: Ia2bbad1cb064825eef7d1847b42bd34189be4dff
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40656
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Shahar Havivi <shavivi@redhat.com>
Continuous-Integration: Jenkins CI