All modules configure should run on upgrade
Currently we run only libvirt and multipath configure. The bugs raises
when we add those calls to newer vdsm  version, and on upgrade the %postun
script of the updated version is ran.

For example, adding the new multipath configure in 3.6 won't persist on upgrade
from 3.5 because the configure call does not exist in 3.5.

With this patch from 3.6 and above we won't need
to worry about other modules - all configurators will run on upgrade.

Change-Id: If6ab88c2dc6321864d9d934355e618f70e95ca10
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39492
Reviewed-by: Yeela Kaplan <ykaplan@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>