password: Prevent password logging and persisting
Vdsm should not expose passwords in log messages. We have code in HSM,
protecting some methods that accept connection lists, which may include
a password parameter. However this is not the right place to protect
passwords, as they are logged now by the jsonrpc server.

This patch introduces the password library, providing:

  ProtectedPassword     class for wrapping password value, so it is not
                        logged or persisted by mistake.

  protect_passwords     replace passwords value with ProtectedPassword
                        object in nested structures.

  unprotect_passwords   replace ProtectedPassword objects with the
                        actual password value.

The jsonprc server protect passwords in request parameters, and
unprotect passwords in response result.  This is safest and simplest
way, as we cannot forget to protect a method.

The xmlrpc server handle protecting and unprotecting in the method
level, since this is the only place where we can detect a password
parameter.

Passwords read from iscsi session and libvirt password file are also
protected when they enter the application.

Code that needs to access protected password value must access the
password.value attribute.

Note that xmlrpc server has incomplete and fragile password protection
for non-irs methods (see wrapApiMethod). Fixing this needs more work.

Change-Id: Icc849ad8bdc1b1fd09884e18038427d96e66110f
Bug-Url: https://bugzilla.redhat.com/1220039
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40707
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI