Revert "getRoute verb added"
getRoute verb was introduced due to our mistake. It is not useful
for its task, was never used by Engine, and was never part of a released
version of vdsm. Instead of using it, Engine is to perform DNS resolution
of the host name and compare it to the addresses found on the interfaces
reported by getVdsCapabilities.

This reverts commit 41a839fc1d7d2e30262b961e33bf25f733a8de02.

Change-Id: Ifd1fa1c9a0af5218035affe8b41e4e24fa8202e6
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40749
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI