ceph: Add option to clear other secrets
Typically engine will add new secrets when activating a Cinder/Ceph
storage domain or when a user add new secret to an active domain. Engine
will remove secrets when deactivating a domain, moving the host to
maintenance, or when a user delete a secret form an active domain.

When host is unreachable when engine try to unregister secrets, the host
will contain stale or unneeded secrets. When the host becomes UP again,
engine would like to register all the secrets required for the current
active storage domains, and clear any other secret.

Change-Id: I476e5aa0bd4a48d1837de09c59597f180fa38823
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40853
Tested-by: Daniel Erez <derez@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI