resize-lun: Resize devices in connectStorageServer
While a host was not connected to a storage server, the underlying
devices backing up a multipath devices may have been resized.

To ensure host is synced with the devices state, we refresh storage,
rescanning devices and resizing them if needed when the host connects to
a storage server.

Change-Id: Ia574a59047d2fb5dd3ac7360c2c3a9dde05237ef
Bug-Url:https://bugzilla.redhat.com/609689
Signed-off-by: Fred Rolland <frolland@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40864
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Tested-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Jenkins CI
Continuous-Integration: Jenkins CI
Tested-by: Jenkins CI
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>