network: allow custom bondOption
Allow 'custom' in bond options. This value will be accepted by
Bond.validateOptions method but it will not be passed to configurator.

Thanks to this patch, users will be allowed to pass custom properties
to setupNetworks hooks.

This path was reverted in past because of malfunction, now it should
work, there is a new test for sure.

Change-Id: I9ace532d959bc3195a8a92b4536bdc0062bc7d1d
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1234867
Reviewed-on: https://gerrit.ovirt.org/40882
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>