gluster: raise proper execption if the given mount point is in use.
Currently if an already used mount point is given for the brick
create function, it does all the lvm tasks and returns Unexpected
Exception when it tries to mout the device at the end of brick
create scenario. This patch will check the availablity of the given
mount point and returns an appropriate error message.

Change-Id: Ic12acd2e4bf6a88936af582bb17d50d72cdfb84f
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1209475
Signed-off-by: Timothy Asir Jeyasingh <tjeyasin@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40892
Continuous-Integration: Jenkins CI
Reviewed-by: Bala.FA <barumuga@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>