vmstats: do not catch unlikely ZeroDivisionError
Nowadays, the only foreseeable reason to have ZeroDivisionError
in the vmstats functions, is when interval is zero.

This case is in turn very very unlikely, as it should
happen only in the extravagant case of two polling cycles racing
together, saving bad bugs in the sampling code.

So, drop the ZeroDivisionError from the handled
exception, and we add an explicit guard to be sure
we record these end-of-the-world-like cases.

Change-Id: I6c6719414d185a8b359b5d8a9f4a28ac87b64327
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40916
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>