migration: remove getVmStats before migration
Calling getVmStats before migration to check if the
same vm exists on the destination host is redundant.
If it does, then our system is in much bigger trouble
and migration will fail anyway.
and if the connection was not established we will fail
in the next step of the migration anyway.

Change-Id: Ic83ab46173d5d7e5bfd85bf00b8f86a530da5dd4
Signed-off-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39202
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Feng Yang <yangfeng@cloud-times.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>