ceph: Undefine existing secret if usage changed
When re-defining a secret, libvirt does not allow to change the usage
id.  We know that we will have to change the usage id in the next
version, when moving secrets from Cinder provider to the storage domain
using the Cinder provider.

So now lookup existing secrets and undefine secrets with different
usage id before defining new secret.

Change-Id: Ib258f0be1cdd4f2672b9adaad35da13d48279ce6
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40967
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Daniel Erez <derez@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>