ceph: Clear all secrets during startup and shutdown
Libvirt secrets are needed only when vdsm is connected to engine. To
ensure that we don't leave stale secrets or secrets which are not needed
by the active storage domains, we clear all secrets when the service
starting or stopping. Engine will register the required secrets when
connecting to the host or when activating or deactivating storage
domains.

Change-Id: I5c5aa36ed752642745b7febc785750b438c102bb
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40968
Reviewed-by: Daniel Erez <derez@redhat.com>
Tested-by: Daniel Erez <derez@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>