gluster: Adds a new element 'additionalFeatures' to getVdsCaps.
This patch adds a new element 'additionalFeatures' to getVdsCaps,
which gives the list of additional fetures supported by the host.
This patch also adds additional gluster features supported in vdsm
to the above list, if the host has vdsm support for features like
snapshot, geo-replication and disk-provisoning they will be added
to the list.

Change-Id: Ic203fa0565914d528ae457427f6f1ff7708017d2
Signed-off-by: Darshan N <dnarayan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40044
Reviewed-by: Bala.FA <barumuga@redhat.com>
Continuous-Integration: Bala.FA <barumuga@redhat.com>