multiapth: Remove unsupported getuid_callout option
This option is not supported since EL 7, and cause a warning when
running multipath command line tool.

This option was required on EL 6, so we kept it in our configuration to
simplify the code. Now that we do not support EL 6, we can remove this
and simplify the code.

Change-Id: If1059bfe2e803e304c51831a4063d21b2af3abd6
Bug-Url: https://bugzilla.redhat.com/1172186
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/41241
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI