gluster:fix NoneType error for some devices does not contain sysfsPath
Currently devices list returns the list of available disks and
its fs type properly. However it fails for some devices which does
not contain sysfsPath and the dev details become None.

This is a workaround patch. This fixes this error by assigning empty
dictonary if the dev is empty or None. This will be removed once
the dependent bugs fixed

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1225417
Change-Id: I54efdf3958519e2ae038b527a03e0b57569c2e8f
Signed-off-by: Timothy Asir Jeyasingh <tjeyasin@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/41262
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Ramesh N <rnachimu@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/41466
Tested-by: Ramesh N <rnachimu@redhat.com>
Reviewed-by: Bala.FA <barumuga@redhat.com>
Continuous-Integration: Bala.FA <barumuga@redhat.com>