net tests: skip test_setupNetworks_on_external_bond
commit f80e6162f introduced test_setupNetworks_on_external_bond that
depends on a subtle difference in the semantics of restoreNetworks that
depends on the persistence model. With unified persistence, the
restoration re-creates the network structure in kernel. However, with
ifcfg persistence, only the configuration files are written to disk.

Until this difference is solved, this patch skips the test on when run
with ifcfg persistence.

Bug-Url: https://bugzilla.redhat.com/1223457
Change-Id: I9d34e316ef7c56b0d8f8aa019595341e34646dc6
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/41338
Reviewed-by: Ondřej Svoboda <osvoboda@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-on: https://gerrit.ovirt.org/41367
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>