tests: Wait until bridge is down before deleting it
Previous code was using evil time.sleep(0.5), which tends to break on our
overloaded slaves with this error:

    RuntimeError: ("Command ['/usr/sbin/brctl', 'delbr', 'vdsmtest-QdYgr']
    returned non-zero exit status 1.", '', "bridge vdsmtest-QdYgr is still up;
    can't delete it\n")

Now we wait up to 2 seconds for a proper 'down' netlink event. Before this
change running tcTests.py took about 8.7 seconds. With this change, it takes
only about 0.7 seconds. Hopefully this will also fix the random failures we see
on the CI.

Change-Id: I0e4b82fe488356c7553b2243dae21cab39a56939
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/41362
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Continuous-Integration: Jenkins CI