vdsm-tool: is-configured: on el6, we must disable sysv's libvirtd job
In the old libvirt_configure.sh.in script we used to run chkconfig
libvirtd off. It has been missing from the new configurators/libvirt.py
implementation. This patch adds that part which introduced a regression
after first host restarts. On upgrade, we check is-configured for
libvirt module - if sysv still handles the service it will report
NOT_CONFIGURED and reconfigure it to use upstart instead of sysv.

Label: ovirt-3.5-only - initctl hack is relevant only to el6
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1222154
Change-Id: I6326a34cc9e4bee82a427f2543354438205cd11d
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/41144
Reviewed-by: Dima Kuznetsov <dkuznets@redhat.com>
Continuous-Integration: Jenkins CI