virt: Do not perform disconnect actions when destroyed
Up until now, we did blindly try to perform actions like desktopLock
on the guest agent, even if the VM was already destroyed.
This can cause tracebacks to be present in the logs since we won't
be connected to the guest agent channel anymore.

Example of traceback:

Traceback (most recent call last):
  File "/usr/share/vdsm/virt/guestagent.py", line 357, in desktopLock
    self._forward("lock-screen")
  File "/usr/share/vdsm/virt/guestagent.py", line 227, in _forward
    self._sock.send(message)
  File "/usr/lib64/python2.7/socket.py", line 170, in _dummy
    raise error(EBADF, 'Bad file descriptor')
error: [Errno 9] Bad file descriptor

This patch now at least checks that '_destroyed' is not 'True'.
This will not solve this for all the cases, since there is still
room for a race condition however should improve the times of this
happening significantly.

Change-Id: I34d384a057e2d6b9093b5ad9ce67d0ba33c42264
Bug-Url: https://bugzilla.redhat.com/1197460
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/41373
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>