virt: allow a single dev from IOMMU group to be passed through
This patch introduces new spec param for hostdev, indicating that the
device should not be passed to host, but only detached and reattached.

When assigning host device, engine still needs to send the whole group
in vmCreate command, but devices that should not be present need to
have this tag set.

Change-Id: Ic266e74b5ffc4ae5c4627b004ade55a19c969511
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/41508
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>