gluster: Added an optional argument to mountBrokerSetup verb
Added an optional argument called "partial" to mountBrokerSetup
verb. This argument takes boolean value. If true only mount broker
root creation is done as part of mountBrokerSetup, else along with
root creation user addition and few mount broker options are also
set.
This is introduced because in only one of the slave node's all the
operations mentioned above has to be done, this is synced to all
the remaining nodes automatically. Only mount broker root creation
has to be performed on all the slave nodes. Engine can call this
verb with partial as false on one of the nodes and on all the other
nodes verb can be called with partial set to true.

Change-Id: I732675c611221cf10f191f69990f78bd713da245
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1220345
Signed-off-by: Darshan N <dnarayan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/41600
Reviewed-by: Sahina Bose <sabose@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Bala.FA <barumuga@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI