net: bonding: validate options based on pre-dumped ones
The removed code created a bond device only to check if the provided
options are recongnized by the kernel. This was buggy, as the options
depends on the bonding mode. It was also redundant, as we already depend
on pre-dumped maps of bond options per bond mode, so we can (and should)
use it here, too.

Change-Id: I03fce903280d5eda817e713a51c8079dab9d7649
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/41616
Continuous-Integration: Jenkins CI
Reviewed-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-by: Petr Horáček <phoracek@redhat.com>