StorageDomainManifest: introduce class heirarchy
Many of the functions in the StorageDomain class family serve to provide
information about the storage domain and its contents (ie. images and
volumes).  Other functions manipulate image or volume metadata in a
fine-grained, reusable way (ie. extendVolume).  As we refactor the
storage code to operate without an SPM the new code will benefit from
reusing these Storage Domain "manifest" functions.

In order to make reuse clean and safe, we want to refactor the
StorageDomain class by splitting the reusable elements into a
StorageDomainManifest class.  The StorageDomain object will initialize
and use a StorageDomainManifest object for much of its normal
operations.  Meanwhile, the new SDM code may use a StorageDomainManifest
object directly without having access to the non-sharable parts of
StorageDomain.

In an effort to reduce churn during the refactoring, when moving a
method into its new class, I will leave behind a stub in the original
class which simply passes the call through to the method's new location.
Once all code has been refactored, I'll remove stubs for any methods
which are not currently accessed outside of the class.

This patch removes volumeTests.FileVolumeGetVSizeTest since it is not
really a Volume test and it would be broken by this patch anyway.  It
will be restored as a storageDomainManifestTest in a followup patch.

Change-Id: Ia275424a8d423722efb229a4d8545f035fe5fa51
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/41993
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>