tests: Do not assume that storage domain properties are writable
We plan to make storage domain properties readonly as part of sdm
refactoring. In fileSDTests and volumeTests we assumed that subclassing
StorageDomain and setting self.sdUUID and self.mountpoint is possible,
but this assumption is wrong, since the super class can change the
implementation.

Now we use properties on the testing subclass, so it works even if the
super class made the properties readonly.

Change-Id: Icc589a118e9062ec38872880496b466a7c429b5a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42006
Continuous-Integration: Jenkins CI
Reviewed-by: Adam Litke <alitke@redhat.com>