net: reset bonding options to defaults.
When an existing bond is edited, replacing bonding options in the
ifcfg file and ifdown+ifup is not enough. This is because ifcfg
script will only set the options that are indicated. This is a
bug since what is persisted is not what is eventually configured on
the edited bond.
This is solved by resetting the non default options back before ifup.
testSetupNetworksResizeBond was changed to point at the bug and prove
it is solved.

Change-Id: Ibe720674e63119e954265735e9f5ccb994c3531a
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1203422
Reviewed-on: https://gerrit.ovirt.org/42115
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-on: https://gerrit.ovirt.org/42958