caps: workaround for libvirt's canonical name in machine types
Libvirt is supposed to return the machine type as 'text' attribute of the
XML. For the default machine type there is an alias in the "canonical"
attribute. Still each machine type is a separate <machine> tag.
In EL7 there is a regression in libvirt that the proper machine type's
name is not reported, only the "canonical" alias is - as part of the
default machine type entry.

This patch is workaround so we don't miss the latest machine type
and can be dropped once libvirt is fixed (bug 1229666).

Change-Id: Ic0bade940d7d548be18eb779cde324e37231ce73
Bug-Url: https://bugzilla.redhat.com/1229396
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42082
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>