gluster: refresh lvm devices after brick create
Issue:
Brick create function in vdsm gluster module which creates lvm
devices like pv, vg, thinlv and format the lv and mount it
using python-blivet. In the brick creation process, the created
vgs are not visible to lvm vgs command and the lvm devices and
its details are not available or not listed by devices list
function.

Cause:
On every successful brick creation the thin lv mount details
will be updated into fstab entry to get the devices mounted
automatically on reboot.

Because of this visibility issue with lvm devices, the thin lvs
which are mounted during brick create process are not getting
mounted automatically on any number of reboot. The Blocking bug
filed in https://bugzilla.redhat.com/show_bug.cgi?id=1227788

Observations:
This issue is observed in any plain RHEL6.6 or RHEL6.7 node
with the latest kernel (2.6.32-567.el6.x86_64) and python-
blivet installed with or without vdsm installed in the node.
This issue identified in the rhel system where it has the
default lvm configuration.

Findings:
According to LVM2 man page documentation, vgscan should
take place automatically and the manuel vgscan required only
when there is any hardware change. The created vg details should
be updated automatically. But it fails here some how. This shows
that, this is absolutely an lvm issue. We have to run pvscan
or vgscan to make lvm see vgs or lvs. Bug for lvm filed in
https://bugzilla.redhat.com/show_bug.cgi?id=1230495
However there may be a little chances for blivet could cause
this issue by passing some causing arguments to lvm. Blivet team
is working on this to verify this.

This issue never observed on any reboots or brick creates;
when we execute a lvm pvs command or vgscan command at-least
once afer any first brick create.

Fix:
This patch refresh lvm devices after brick create
by running vgscan. This is a workaround patch and this
will be removed once the lvm bz#1230495 is fixed.

Kindly refer the bugs for any further details or logs.

Change-Id: I6a6a10a2746c6f44a16de41c153ce46460d37b7f
Signed-off-by: Timothy Asir Jeyasingh <tjeyasin@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1227742
Bug-Url: https://bugzilla.redhat.com/1227788
Signed-off-by: Timothy Asir Jeyasingh <tjeyasin@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42142
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42365
Tested-by: Darshan N <dnarayan@redhat.com>
Reviewed-by: Bala.FA <barumuga@redhat.com>
Continuous-Integration: Bala.FA <barumuga@redhat.com>