net: during boot, restore only networks that were actually changed.
This is an optimization that will shorten boot time when there are many
networks. If also, all persisted networks will be persisted with onboot=true,
then in most case, all the hard of restorattion will be done by network
service and vdsm will only interfere to fix the changed networks and bonds.

Change-Id: I66307a7cf6b0f7269a1ae04aa7c1748dde768144
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1203422
Reviewed-on: https://gerrit.ovirt.org/42355
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI