net: Drop default bonding options
On old engines, no bonding options were provided to VDSM and it
used to configure it's own default options. Nowadays, those options
are always provided by engine. If a user will setup a network on an
existing bond, VDSM takes ownership of the bond and configures the
options on it.
The code for the default options was very local and options were
set 'too late' to be persisted. Hence this code was incorrect and
is not needed anyway.
The change in behavior is that from now on, if a bond with no options
is explicitly requested, the kernel will be asked to create a bond
without any options making the default kernel options effective.

Change-Id: I9e4a8d832891f62d1fcc412606b014fa4f874062
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42594
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>