test: make sure that jsonrpc test timeout
We want to reuse recently created call method with timeout. By using it
we would make sure to timeout when the response won't arrive. Whenever
this situation happens we raise JsonRpcNoResponseError.

When we added timeout method which takes more time to process than the
client wait on it the code close server which is problematic. After slow
method processing finishes and the server side is already closed the is
nice stack trace on the console so we added a check so we won't pollute
the console during test execution.


Change-Id: I51b2cf980c3c51467f7b0bef677d8b2bbdd43bf5
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/42541
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>