virt: Set cloexec flag on channel sockets
To counter act potential leakage of open handles, this patch is setting the
CLOEXEC flag on the sockets used for the guest agent channels.
It is known that this is not totally race free however it will still improve
the number of file handles leaked.

Change-Id: Ic4a452f7837267434bc836fced4c2efd92d745cf
Bug-Url: https://bugzilla.redhat.com/1297414
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42570
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51878
Continuous-Integration: Francesco Romani <fromani@redhat.com>