events: make sure to send only one event per status
Initially we made an assumption that engine can handle more than one
event containing the same status. The assumption is not true and there
are flows where engine logic assumes only one status change.

One of the problematic flows is vm destroy for which having duplicate
means that the operation failed and recovery flow is triggered.

Change-Id: I044a8f409ccdac4210784e21167bab320ddfe808
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/42579
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>