resize-lun: Fix resizing when vg has multiple pvs
When resizing a PV, LVM needs the other PVs that are part of the VG
in the filters parameters. After resizing the PV, LVM has to update the
metadata of the VG. This operation will fail if not all the PVs that are
part of the VG are included in the filter parameters.

Change-Id: I4213f82a0307dc91202078b53c0b2c2f24433899
Bug-Url:https://bugzilla.redhat.com/609689
Signed-off-by: Fred Rolland <frolland@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42608
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>