net: tests: wait for a specific state of a device
5d6c571d added _waitForKnownOperstate. This improved tests
reliability but it is not enough. It may happen that a test will
fail because the initial device state is down and inside a
nonChangingOperstate context manager, it changes to up. It is safer
to ensure that the initial operation state is not only other than
'unknown' but a specific one (most probably 'up').
Also, the state names where replaced with constants in
lib/vdsm/netinfo.py for future use.

Change-Id: Icbcfe6423894b2b8806c355dd093ddfd66eb75f0
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42675
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI