virt: add logic for POWER cpu xml element
Current flow in appendCpu caused ppc to be skipped, generating empty
element (<cpu/>). In order to specify cpu model on ppc, we have to use
libvirt's cpu with POWER8 model.

This is currently the only way of safely adding CPU to ppc64le host,
and hopefully will be enhanced in future.

Change-Id: I9303b76904ef1344508136343104b37c09d2a370
Bug-Url: https://bugzilla.redhat.com/1259318
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42736
Continuous-Integration: Jenkins CI
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>