lib: migration: add response.is_error helper
Add helper function to make the code more explicit
and self-documenting.
The is_error function will raise the new MalformedResponse
exception if the response value is missing one of
the mandatory keys.

This is new behaviour, but please note that previously
the code either (rarely) checked for the presence of
the keys, so no breakage, or just exploded with less
clear KeyError.

Change-Id: I6ccefec4f1bebcb2ca64f0bc9f6b9e9954dbf04c
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42795
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>