net: tests: do not use KernelConfig.__eq__ in tests
Until now, functional tests used the __eq__ to assert that a
KernelConfig instance is equal to a RunningConfig instance. This
hides the normalization process inside __eq__ and makes the life of
the tester very hard if something really breaks. We trade some magic
here for a better error message.

Change-Id: I9b142d1b62e293bf094f650c3c691d65f684231d
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42813
Continuous-Integration: Jenkins CI
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>