Live Merge: Prevent merge when base volume is too small
When a disk with snapshots is resized you can end up with an image chain
containing volumes of different sizes.  For example: a VM may start with
volume A, a snapshot operation causes volume B to be added to the chain,
and a resize operation enlarges volume B.  This mixed-size chain
presents a special case for live merge.

Libvirt's virDomainBlockCommit behaves differently depending on whether
the base volume is on block storage or file storage and whether
the base volume is raw or qcow2:

File volume, cow format:
 - Update qcow header
File volume, raw format:
 - Truncate volume file to the new size
Block volume, cow format:
 - Update qcow header
Block volume, raw format:
 - Fail because qemu cannot extend a raw block device

Since we know the raw block case is not possible we can detect this and
return a special error message to engine.  To avoid this error, engine
should use the SPM host to extend the raw volume to the correct size
before starting the merge.

We don't _have_ to change vdsm since the current code fails gracefully
(albeit with a traceback in the logs).  If engine always performs the
resize operation before calling merge we'll never see this error.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1232481
Change-Id: Ibf77a7c5108b500c6ec34653ef7570a841def1b4
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42836
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>