net: use KernelConfig instead of RunningConfig in setupNetworks
Using RunningConfig is not enough because it is not necessarily
equal to what is really configured in the kernel. Also, it might
not be present at all (during boot). since this code runs only if
the network is in _netinfo.networks, it is also safer to use
KernelConfig.
Note that just as in vdsm-restore-net-config, a KernelConfig instance
and a {Persistent,Running}Config instance cannot be directly compared
without normalizing the latter. This is because a few language
differences and a few default values used in the latter.

Change-Id: I41232de5b079e6192b3c0f031d0cb24036583c43
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42903
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>