tests: Add RWLock tests
Current RWLock looks too complex and has no tests. This patch fixes the
later issue, so we can safely simplify the implementation later.

Only one statement is not covered by the tests, handling of Queue.Full -
this can never happen since we use unbounded queue. This code should
be removed.

The new tests reveal that lock demotion is broken; when replacing an
exclusive lock with shared lock, we actually continue to hold an
exclusive lock until we release the lock. This test is marked as broken
so to avoid breaking the build.  I think we don't use lock demotion in
the current code and this should be fixed by not supporting demotion.

Change-Id: If404f0e8c68fcdb2f7643bdd6d5c1f97f230a227
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42907
Continuous-Integration: Jenkins CI
Reviewed-by: Adam Litke <alitke@redhat.com>