Live merge: Update base size after live merge
When performing a live merge, data is copied from a top volume into a
base volume.  If the top volume is larger than the base volume (which
can happen if the drive size was extended), libvirt will change the size
of the base volume to match that of the top volume.  When synchronizing
metadata after the merge, we need to update the 'capacity' field of the
base volume to reflect the new size.  We do this inside the
LiveMergeCleanupThread to ensure that it gets retried in the event of
storage connection problems or vdsm restarts.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1251958
Change-Id: Iae354de36db63ae3bf4b4fc7f72df5e306035784
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42921
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45950
Continuous-Integration: Francesco Romani <fromani@redhat.com>